requirements.txt
python-telegram-bot==21.0.1
SQLAlchemy==2.0.29
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1
psycopg2-binary
selenium==4.21.0
beautifulsoup4==4.12.3
lxml==5.2.1